Create MMT DIRECTOR BETRAYS NPP
The deputy managing director of MMTL Osei Richard Bamfo popularly known as Nana Osei has switched camp. Mr Bamfo came into the political race in 2016 where he contested for the parliamentary candidates position on the ticket of the  ruling New Patriotic Party but lost.  The Nana Addo government found him a space to compensate him in his government by appointing him as DMD GeForce Metro mass transit and has since been at post.
On Sunday 12th June,2022, Mr Bamfo met a group of delegates and made it clear to them on his intentions to contest come 2024 but this time on the ticket of United Progressive Party(UPP).
He added that the NPP govt has failed Ghanaians and used Metro mass as an example stating that the government has neglected all state owned institutions and making it difficult for the state owned institutions to operate.
He said the government has focused on policies that he feels can benefit his "babies with sharp teeth" just to help him recoup all the monies he spent to become the president. Mr Vamfo stated emphatically that it was part of the reasons the president was so keen on ensuring the victory of the Ashanti regional Chairman "WONTUMI" to continue his grab all agenda.